requirements.txt
appdirs
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
numpy
packaging
protobuf
pyparsing
six
tensorflow
Werkzeug
scipy
Pillow
keras
h5py
imageio
scikit-image